# Sistema_Gestión_Donante/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema de Gestión de Donantes - API",
        "description": "API para el módulo de donantes del sistema de adopciones de animales",
        "contact": {
            "email": "admin@sistema-adopciones.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/api/informacion-contactos": {
            "get": {
                "tags": [
                    "Información Contacto"
                ],
                "summary": "Obtener toda la información de contactos",
                "operationId": "f34280f03add4e0b866933160f7da624",
                "responses": {
                    "200": {
                        "description": "Lista de información de contactos obtenida exitosamente"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Información Contacto"
                ],
                "summary": "Crear nueva información de contacto",
                "operationId": "642b4583200d360a07296b2f2f46c812",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "donante_id",
                                    "nombre_contacto",
                                    "telefono",
                                    "relacion"
                                ],
                                "properties": {
                                    "donante_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nombre_contacto": {
                                        "type": "string",
                                        "example": "Pedro García"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "555-0201"
                                    },
                                    "relacion": {
                                        "type": "string",
                                        "example": "Hermano"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Información de contacto creada exitosamente"
                    },
                    "400": {
                        "description": "Datos inválidos"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            }
        },
        "/api/informacion-contactos/{id}": {
            "get": {
                "tags": [
                    "Información Contacto"
                ],
                "summary": "Obtener información de contacto por ID",
                "operationId": "ef8b5b3a89e43787af8dc4af4b9975f5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la información de contacto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Información de contacto obtenida exitosamente"
                    },
                    "404": {
                        "description": "Información de contacto no encontrada"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            },
            "put": {
                "tags": [
                    "Información Contacto"
                ],
                "summary": "Actualizar información de contacto existente",
                "operationId": "560d0a83b0ca339cf073ae81b7a717f7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la información de contacto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "donante_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nombre_contacto": {
                                        "type": "string",
                                        "example": "Pedro García Actualizado"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "555-9999"
                                    },
                                    "relacion": {
                                        "type": "string",
                                        "example": "Primo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Información de contacto actualizada exitosamente"
                    },
                    "400": {
                        "description": "Datos inválidos"
                    },
                    "404": {
                        "description": "Información de contacto no encontrada"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Información Contacto"
                ],
                "summary": "Eliminar información de contacto",
                "operationId": "9f4373159ffe57217444b67669164f17",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la información de contacto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Información de contacto eliminada exitosamente"
                    },
                    "404": {
                        "description": "Información de contacto no encontrada"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            }
        },
        "/api/informacion-contactos/donante/{donanteId}": {
            "get": {
                "tags": [
                    "Información Contacto"
                ],
                "summary": "Obtener información de contacto por donante",
                "operationId": "42c2a296881fb51d85c7815b81590f70",
                "parameters": [
                    {
                        "name": "donanteId",
                        "in": "path",
                        "description": "ID del donante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de información de contacto por donante obtenida exitosamente"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            }
        },
...
    },
...
}